+++ client/src/main.ts
@@ -4,4 +4,4 @@ import { AppComponent } from './app/app.component';
 import { appConfig } from './app/app.config';
 
 bootstrapApplication(AppComponent, appConfig)
-  .catch((err: any) => console.error(err));
+  .catch((err: unknown) => console.error(err));